refactor dataset code organization
